requirements.txt
tensorflow==2.18.0
streamlit==1.16.0
Pillow==8.3.2
